Split NDVI children into separate module
