+++ src/aatoolbox/datasources/usgs/usgs_ndvi.py
@@ -767,183 +767,3 @@ class _UsgsNdvi(DataSource):
         # find two groups, first for year second for dekad
         regex = re.compile(r"(\d{4})_(\d{2})")
         return [int(x) for x in regex.findall(filename)[0]]
-
-
-class UsgsNdviSmoothed(_UsgsNdvi):
-    """Base class to retrieve smoothed NDVI data.
-
-    The retrieved data is the smoothed NDVI values
-    processed by the USGS. Temporal smoothing is done
-    to adjust for cloud cover and other errors.
-    Data for the 3 most recent dekads is not fully
-    smoothed, and are re-smoothed at the end of the
-    3 dekad period.
-
-    Parameters
-    ----------
-    country_config : CountryConfig
-        Country configuration
-    start_date : Union[date, str, Tuple[int, int], None]
-        Start date. Can be passed as a ``datetime.date``
-        object and the relevant dekad will be determined,
-        as a date string in ISO8601 format, or as a
-        year-dekad tuple, i.e. (2020, 1).
-    end_date : Union[date, str, Tuple[int, int], None]
-        End date. Can be passed as a ``datetime.date``
-        object and the relevant dekad will be determined,
-        as a date string in ISO8601 format, or as a
-        year-dekad tuple, i.e. (2020, 1).
-    processed_file_suffix : str
-        Suffix for processed file.
-    """
-
-    def __init__(
-        self,
-        country_config: CountryConfig,
-        start_date: Union[date, str, Tuple[int, int], None] = None,
-        end_date: Union[date, str, Tuple[int, int], None] = None,
-        processed_file_suffix: str = "",
-    ):
-        super().__init__(
-            country_config=country_config,
-            data_variable="smoothed",
-            data_variable_suffix="",
-            data_variable_url="temporallysmoothedndvi",
-            start_date=start_date,
-            end_date=end_date,
-            processed_file_suffix=processed_file_suffix,
-        )
-
-
-class UsgsNdviPctMedian(_UsgsNdvi):
-    """Base class to retrieve % of median NDVI.
-
-    The retrieved data is the percent of median NDVI
-    values calculated from 2003 - 2017, as
-    processed by the USGS.
-
-    Parameters
-    ----------
-    country_config : CountryConfig
-        Country configuration
-    start_date : Union[date, str, Tuple[int, int], None]
-        Start date. Can be passed as a ``datetime.date``
-        object and the relevant dekad will be determined,
-        as a date string in ISO8601 format, or as a
-        year-dekad tuple, i.e. (2020, 1).
-    end_date : Union[date, str, Tuple[int, int], None]
-        End date. Can be passed as a ``datetime.date``
-        object and the relevant dekad will be determined,
-        as a date string in ISO8601 format, or as a
-        year-dekad tuple, i.e. (2020, 1).
-    processed_file_suffix : str
-        Suffix for processed file.
-    """
-
-    def __init__(
-        self,
-        country_config: CountryConfig,
-        start_date: Union[date, str, Tuple[int, int], None] = None,
-        end_date: Union[date, str, Tuple[int, int], None] = None,
-        processed_file_suffix: str = "",
-    ):
-        super().__init__(
-            country_config=country_config,
-            data_variable="percent_median",
-            data_variable_suffix="pct",
-            data_variable_url="percentofmedian",
-            start_date=start_date,
-            end_date=end_date,
-            processed_file_suffix=processed_file_suffix,
-        )
-
-
-class UsgsNdviMedianAnomaly(_UsgsNdvi):
-    """Base class to retrieve NDVI anomaly data.
-
-    The retrieved data is NDVI anomaly data calculated
-    as a subtraction of the median value from the
-    current value. Negative values indicate less
-    vegetation than the median, positive values indicate
-    more vegetation.
-
-    Parameters
-    ----------
-    country_config : CountryConfig
-        Country configuration
-    start_date : Union[date, str, Tuple[int, int], None]
-        Start date. Can be passed as a ``datetime.date``
-        object and the relevant dekad will be determined,
-        as a date string in ISO8601 format, or as a
-        year-dekad tuple, i.e. (2020, 1).
-    end_date : Union[date, str, Tuple[int, int], None]
-        End date. Can be passed as a ``datetime.date``
-        object and the relevant dekad will be determined,
-        as a date string in ISO8601 format, or as a
-        year-dekad tuple, i.e. (2020, 1).
-    processed_file_suffix : str
-        Suffix for processed file.
-    """
-
-    def __init__(
-        self,
-        country_config: CountryConfig,
-        start_date: Union[date, str, Tuple[int, int], None] = None,
-        end_date: Union[date, str, Tuple[int, int], None] = None,
-        processed_file_suffix: str = "",
-    ):
-        super().__init__(
-            country_config=country_config,
-            data_variable="median_anomaly",
-            data_variable_suffix="stmdn",
-            data_variable_url="mediananomaly",
-            start_date=start_date,
-            end_date=end_date,
-            processed_file_suffix=processed_file_suffix,
-        )
-
-
-class UsgsNdviYearDifference(_UsgsNdvi):
-    """Base class to retrieve NDVI year difference data.
-
-    The retrieved data is NDVI yearly difference data,
-    calculated as the subtraction of the previous year's
-    NDVI value from the current year's. Negative
-    values indicate the current vegetation is less
-    than the previous year's, positive that there
-    is more vegetation in the current year.
-
-    Parameters
-    ----------
-    country_config : CountryConfig
-        Country configuration
-    start_date : Union[date, str, Tuple[int, int], None]
-        Start date. Can be passed as a ``datetime.date``
-        object and the relevant dekad will be determined,
-        as a date string in ISO8601 format, or as a
-        year-dekad tuple, i.e. (2020, 1).
-    end_date : Union[date, str, Tuple[int, int], None]
-        End date. Can be passed as a ``datetime.date``
-        object and the relevant dekad will be determined,
-        as a date string in ISO8601 format, or as a
-        year-dekad tuple, i.e. (2020, 1).
-    processed_file_suffix : str
-        Suffix for processed file.
-    """
-
-    def __init__(
-        self,
-        country_config: CountryConfig,
-        start_date: Union[date, str, Tuple[int, int], None] = None,
-        end_date: Union[date, str, Tuple[int, int], None] = None,
-        processed_file_suffix: str = "",
-    ):
-        super().__init__(
-            country_config=country_config,
-            data_variable="difference",
-            data_variable_suffix="dif",
-            data_variable_url="differencepreviousyear",
-            start_date=start_date,
-            end_date=end_date,
-            processed_file_suffix=processed_file_suffix,
-        )
